Upload YUI Compressor for Debugging and Testing

+++ scripts/release-0.8.8.sh
@@ -23,6 +23,12 @@ if [ -z "`which python`" ] ; then
 	exit 1
 fi
 
+# test if java is available
+if [ -z "`which java`" ] ; then
+	echo "Java executable not found"
+	exit 1
+fi
+
 # remove directories from previous exports
 rm -rf wcom-trunk
 
@@ -52,18 +58,27 @@ rm -rf database
 rm -rf documentation
 
 # compress js
-svn export https://welcompose.svn.cvsdude.com/welcompose/trunk/scripts/third_party/jsmin.py jsmin.py
-if [ ! -f "jsmin.py" ] ; then
-	echo "JavaScript minifier not found. Download failed, eh?"
+#svn export https://welcompose.svn.cvsdude.com/welcompose/trunk/scripts/third_party/jsmin.py jsmin.py
+#if [ ! -f "jsmin.py" ] ; then
+#	echo "JavaScript minifier not found. Download failed, eh?"
+#	exit 1
+#fi
+
+# compress js
+svn export https://welcompose.svn.cvsdude.com/welcompose/trunk/scripts/third_party/yuicompressor-2.4.2 yuicompressor
+if [ ! -f "yuicompressor" ] ; then
+	echo "JavaScript YUI minifier not found. Download failed, eh?"
 	exit 1
 fi
+
 for file in `find welcompose -type f -name "*.js"` ; do
-	tmpfile="$file"-jsmin
+	tmpfile="$file"-yui
 	cp "$file" "$tmpfile"
-	python jsmin.py < "$tmpfile"  > "$file"
+	java -jar build/yuicompressor.jar -o "$file" "$tmpfile"
 	rm -f "$tmpfile"
 done
-rm -rf jsmin.py
+#rm -rf jsmin.py
+
 
 # compress css
 svn export https://welcompose.svn.cvsdude.com/welcompose/trunk/scripts/third_party/csstidy csstidy
